refactor(mail): keep the split-view geometry with the other constants
splitPane.ts held two Tailwind class strings and no behaviour. constants.ts is
already where that kind of thing lives — FLAGGED_STAR_STYLE is the same shape — so
the module boundary was buying nothing.

The comment explaining why the proportions are shared moves with them, which was
the only part earning its keep.
